feat: Add support for multidimensional arrays
I am using https://github.com/dweiller/zig-strided-arrays.

The only downside of this is, that you need to know the datatype and
dimension of the loaded array/npy file beforehand.
This is fine for my use case but a more dynamic approach might be more
suitable for other cases, where the datatype or dimension is not known
at compile time. I will try to add this in the future. Still learning
comptime and zig idioms.